continued developing floquet TI band calc
Continued calculating TI quasiband structure.

Restricted plotter to -Omega/2, Omega/2 over BZ lines. Found that many
bands need to be calculated to find band structure for large A, since
the first few bands that are calculated are located in quasienergy space
near -A, A -- adding a band moves the branch up by Omega, so many bands
must be calculated to bring the tau_z = + branch to the tau_z = -
branch.